pyproject.toml
[project]
name = "arxiv-papers-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.6.0",
    "comet-ml>=3.49.7",
    "datasets>=3.5.0",
    "evaluate>=0.4.3",
    "huggingface-hub[hf-xet]>=0.30.1",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "tiktoken>=0.9.0",
    "torch>=2.6.0",
    "transformers>=4.50.3",
]
